+++ src/ssimu2/main.hpp
@@ -32,16 +32,16 @@ void memoryorganizer(float3* out, const uint8_t *srcp0, const uint8_t *srcp1, co
 }
 
 template <InputMemType T>
-double ssimu2process(const uint8_t *srcp1[3], const uint8_t *srcp2[3], float3* pinned, int stride, int width, int height, float* gaussiankernel, int maxshared, hipStream_t stream){
+double ssimu2process(const uint8_t *srcp1[3], const uint8_t *srcp2[3], float3* pinned, size_t stride, size_t width, size_t height, float* gaussiankernel, size_t maxshared, hipStream_t stream){
 
-    int wh = width*height;
-    int whs[6] = {wh, ((height-1)/2 + 1)*((width-1)/2 + 1), ((height-1)/4 + 1)*((width-1)/4 + 1), ((height-1)/8 + 1)*((width-1)/8 + 1), ((height-1)/16 + 1)*((width-1)/16 + 1), ((height-1)/32 + 1)*((width-1)/32 + 1)};
-    int whs_integral[7];
+    size_t wh = width*height;
+    size_t whs[6] = {wh, ((height-1)/2 + 1)*((width-1)/2 + 1), ((height-1)/4 + 1)*((width-1)/4 + 1), ((height-1)/8 + 1)*((width-1)/8 + 1), ((height-1)/16 + 1)*((width-1)/16 + 1), ((height-1)/32 + 1)*((width-1)/32 + 1)};
+    size_t whs_integral[7];
     whs_integral[0] = 0;
     for (int i = 0; i < 7; i++){
         whs_integral[i+1] = whs_integral[i] + whs[i];
     }
-    int totalscalesize = whs_integral[6];
+    size_t totalscalesize = whs_integral[6];
 
     //big memory allocation, we will try it multiple time if failed to save when too much threads are used
     hipError_t erralloc;
@@ -77,8 +77,8 @@ double ssimu2process(const uint8_t *srcp1[3], const uint8_t *srcp2[3], float3* p
     rgb_to_linear(src2_d, totalscalesize, stream);
 
     //step 1 : fill the downsample part
-    int nw = width;
-    int nh = height;
+    size_t nw = width;
+    size_t nh = height;
     for (int scale = 1; scale <= 5; scale++){
         downsample(src1_d+whs_integral[scale-1], src1_d+whs_integral[scale], nw, nh, stream);
         nw = (nw -1)/2 + 1;
