+++ src/ssimu2/vapoursynth.hpp
@@ -155,7 +155,7 @@ namespace ssimu2{
         }
     
         d.streamnum = std::min(d.streamnum, infos.numThreads); // vs threads < numStream would make no sense
-        d.streamnum = std::min(d.streamnum, (int)(devattr.totalGlobalMem/(32*4*viref->width*viref->height))); //VRAM overcommit partial protection
+        d.streamnum = std::min((size_t)d.streamnum, (size_t)((size_t)devattr.totalGlobalMem/(32llu*4llu*(size_t)viref->width*(size_t)viref->height))); //VRAM overcommit partial protection
         d.streamnum = std::max(d.streamnum, 1); //at least one stream to not just wait indefinitely
         d.streams = (hipStream_t*)malloc(sizeof(hipStream_t)*d.streamnum);
         for (int i = 0; i < d.streamnum; i++){
